@@ -15,36 +15,37 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~8.2.14",
-    "@angular/core": "~8.2.14",
-    "@angular/forms": "~8.2.14",
-    "@angular/platform-browser": "~8.2.14",
-    "@angular/platform-browser-dynamic": "~8.2.14",
-    "@angular/router": "~8.2.14",
+    "@angular/common": "~9.0.0",
+    "@angular/core": "~9.0.0",
+    "@angular/forms": "~9.0.0",
+    "@angular/platform-browser": "~9.0.0",
+    "@angular/platform-browser-dynamic": "~9.0.0",
+    "@angular/router": "~9.0.0",
     "@capacitor/android": "~1.5.0",
     "@capacitor/core": "~1.5.0",
     "@capacitor/ios": "~1.5.0",
-    "@ionic/angular": "~4.11.10",
+    "@ionic/angular": "~5.0.0-rc.3",
     "@ionic/storage": "~2.2.0",
     "core-js": "~3.6.4",
     "rxjs": "~6.5.4",
+    "tslib": "^1.10.0",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.800.6",
-    "@angular-devkit/build-angular": "~0.800.6",
-    "@angular-devkit/core": "~8.3.25",
-    "@angular-devkit/schematics": "~8.3.25",
-    "@angular/cli": "~8.3.25",
-    "@angular/compiler": "~8.2.14",
-    "@angular/compiler-cli": "~8.2.14",
-    "@angular/language-service": "~8.2.14",
+    "@angular-devkit/architect": "~0.900.1",
+    "@angular-devkit/build-angular": "~0.900.1",
+    "@angular-devkit/core": "~9.0.1",
+    "@angular-devkit/schematics": "~9.0.1",
+    "@angular/cli": "~9.0.1",
+    "@angular/compiler": "~9.0.0",
+    "@angular/compiler-cli": "~9.0.0",
+    "@angular/language-service": "~9.0.0",
     "@capacitor/cli": "~1.5.0",
     "@ionic/angular-toolkit": "~2.1.2",
     "@types/jasmine": "~3.5.3",
     "@types/jasminewd2": "~2.0.8",
-    "@types/node": "~13.7.0",
-    "codelyzer": "~5.2.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.4.1",
@@ -54,8 +55,8 @@
     "karma-jasmine-html-reporter": "~1.5.2",
     "protractor": "~5.4.3",
     "ts-node": "~8.6.2",
-    "tslint": "~5.17.0",
-    "typescript": "~3.5.3"
+    "tslint": "~6.0.0",
+    "typescript": "~3.7.5"
   },
   "description": "An Ionic project"
 }
