@@ -3,14 +3,14 @@ import { PreloadAllModules, RouterModule, Routes } from '@angular/router';
 
 const routes: Routes = [
   { path: '', redirectTo: 'tea-categories', pathMatch: 'full' },
-  { path: 'login', loadChildren: './login/login.module#LoginPageModule' },
+  { path: 'login', loadChildren: () => import('./login/login.module').then(m => m.LoginPageModule) },
   {
     path: 'edit-tea-category',
-    loadChildren: './edit-tea-category/edit-tea-category.module#EditTeaCategoryPageModule'
+    loadChildren: () => import('./edit-tea-category/edit-tea-category.module').then(m => m.EditTeaCategoryPageModule)
   },
   {
     path: 'tea-categories',
-    loadChildren: './tea-categories/tea-categories.module#TeaCategoriesPageModule'
+    loadChildren: () => import('./tea-categories/tea-categories.module').then(m => m.TeaCategoriesPageModule)
   }
 ];
 
